# examples/t1api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RouteRequest": {
        "properties": {
          "automatic_fallback": {
            "default": false,
            "description": "Only meaningful with model_id set: fall through the cascade if the requested model is exhausted, instead of raising MODEL_QUOTA_EXHAUSTED.",
            "title": "Automatic Fallback",
            "type": "boolean"
          },
          "input_tokens": {
            "default": 0,
            "minimum": 0.0,
            "title": "Input Tokens",
            "type": "integer"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Manual model selection. Omit for automatic routing.",
            "title": "Model Id"
          },
          "plan": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional assertion of the caller's plan. The server resolves the real plan from the key's assignment; a mismatch is rejected, never silently accepted.",
            "title": "Plan"
          }
        },
        "title": "RouteRequest",
        "type": "object"
      },
      "RouteResponse": {
        "properties": {
          "cascade_position": {
            "title": "Cascade Position",
            "type": "integer"
          },
          "considered": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Considered",
            "type": "array"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "policy_name": {
            "title": "Policy Name",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "resolved_plan": {
            "default": "",
            "description": "The plan the server actually applied, from the key's assignment.",
            "title": "Resolved Plan",
            "type": "string"
          }
        },
        "required": [
          "model_id",
          "reason",
          "cascade_position",
          "policy_name",
          "considered",
          "request_id"
        ],
        "title": "RouteResponse",
        "type": "object"
      },
      "ServerDetailResponse": {
        "properties": {
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "server": {
            "$ref": "#/components/schemas/ServerItem"
          }
        },
        "required": [
          "server",
          "request_id"
        ],
        "title": "ServerDetailResponse",
        "type": "object"
      },
      "ServerItem": {
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "last_seen": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "registered_by": {
            "title": "Registered By",
            "type": "string"
          },
          "server_id": {
            "title": "Server Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Tags",
            "type": "object"
          },
          "trust_level": {
            "title": "Trust Level",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "number"
          }
        },
        "required": [
          "server_id",
          "name",
          "address",
          "trust_level",
          "status",
          "capabilities",
          "tags",
          "registered_by",
          "created_at",
          "updated_at",
          "last_seen"
        ],
        "title": "ServerItem",
        "type": "object"
      },
      "ServerListResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "servers": {
            "items": {
              "$ref": "#/components/schemas/ServerItem"
            },
            "title": "Servers",
            "type": "array"
          }
        },
        "required": [
          "servers",
          "count",
          "request_id"
        ],
        "title": "ServerListResponse",
        "type": "object"
      },
      "ServerRegisterRequest": {
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "allow_private_address": {
            "default": false,
            "description": "Set True for local/Tailscale addresses.",
            "title": "Allow Private Address",
            "type": "boolean"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "tags": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Tags",
            "type": "object"
          }
        },
        "required": [
          "name",
          "address"
        ],
        "title": "ServerRegisterRequest",
        "type": "object"
      },
      "ServerUpdateRequest": {
        "properties": {
          "capabilities": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capabilities"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "tags": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "trust_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trust Level"
          }
        },
        "title": "ServerUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Controlled gateway into HyperNix-pip. The client requests an operation; the server decides what exists, what's available, and how much is left. See wiki/T1-API.md for the full contract.",
    "title": "HyperNix T1 API",
    "version": "0.71.5b3"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/models/route": {
      "post": {
        "description": "Not in the original endpoint list \u2014 added to give the routing/\nquota-cascade engine (``t1api.routing``) an HTTP surface, since the\nspec describes the engine's behavior in detail but doesn't enumerate\nan endpoint for it. See wiki/T1-API.md#routing.\n\nAutomatic routing (no ``model_id``) walks the caller's plan cascade.\nManual routing (``model_id`` set) enforces the exhausted-model rule\nexactly as the spec describes: ``MODEL_QUOTA_EXHAUSTED`` unless\n``automatic_fallback=True``.\n\n**Beta 3 change, deliberately breaking Beta 2's shape:** the plan is\nno longer read from the request body. It is resolved from the key's\nserver-side assignment (:meth:`t1api.keys.KeyDirectory.resolve_plan`),\nand a ``plan`` in the body is treated as an assertion \u2014 matching is\nfine, mismatching is refused. Beta 2 let a client name its own plan,\nwhich is the exact thing the design principle forbids: \"the server\ndetermines... which models the user can access, which limits apply,\nwhich fallback models are allowed.\"\n\nA key narrowed to a model subset (``allowed_models`` on its\nassignment) is also checked here, so manual selection can't reach a\nmodel the assignment excludes even when the registry and quota would\nboth allow it.",
        "operationId": "route_model_models_route_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Route Model",
        "tags": [
          "models"
        ]
      }
    },
...
    "/servers": {
      "get": {
        "operationId": "list_servers_servers_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Servers",
        "tags": [
          "servers"
        ]
      }
    },
    "/servers/register": {
      "post": {
        "operationId": "register_server_servers_register_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Server",
        "tags": [
          "servers"
        ]
      }
    },
    "/servers/{server_id}": {
      "delete": {
        "description": "Admin-only and destructive: needs ``?confirm=true``.\n\nDeregistering a server orphans every module deployment that pointed\nat it, so it gets the same explicit-confirmation treatment as\ndeleting a module.",
        "operationId": "delete_server_servers__server_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "title": "Server Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Server",
        "tags": [
          "servers"
        ]
      },
      "patch": {
        "description": "Admin-only: trust-level changes are a security-relevant escalation\n(\"Verify remote server trust before remote uploads\" only means\nanything if promoting a server to trusted is itself gated).",
        "operationId": "update_server_servers__server_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "title": "Server Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Server",
        "tags": [
          "servers"
        ]
      }
    },
...
  }
}